+++ gDRcore/tests/testthat/test-finalMAE_combo_2dose_nonoise.R
@@ -2,8 +2,10 @@ test_that("combo_2dose_nonoise: test_synthetic_data", {
   data <- "finalMAE_combo_2dose_nonoise.RDS"
   original <- get_synthetic_data(data)
   
-  mae <-
-    gDRtestData::generateComboNoNoiseData(cell_lines, drugs, e_inf, ec50, hill_coef)
+  mae <- purrr::quietly(gDRtestData::generateComboNoNoiseData)(
+    cell_lines, drugs, e_inf, ec50, hill_coef, FALSE
+  )
+  expect_length(mae$warnings, 6)
   
-  test_synthetic_data(original, mae, data)
+  test_synthetic_data(original, mae$result, data)
 })