+++ gDRcore/tests/testthat/test-finalMAE_combo_2dose_nonoise3.R
@@ -2,8 +2,10 @@ test_that("combo_2dose_nonoise3: test_synthetic_data", {
   data <- "finalMAE_combo_2dose_nonoise3.RDS"
   original <- get_synthetic_data(data)
   
-  mae <-
-    gDRtestData::generateComboNoNoiseData3(cell_lines, drugs, e_inf, ec50, hill_coef)
-  
-  test_synthetic_data(original, mae, data)
+  mae <- purrr::quietly(gDRtestData::generateComboNoNoiseData3)(
+    cell_lines, drugs, e_inf, ec50, hill_coef, FALSE
+  )
+  expect_length(mae$warnings, 6)
+
+  test_synthetic_data(original, mae$result, data)
 })