+++ gDRcore/tests/testthat/test-finalMAE_combo_matrix_small.R
@@ -1,8 +1,10 @@
 test_that("combo_matrix_small: test_synthetic_data", {
   data <- "finalMAE_combo_matrix_small.RDS"
   original <- get_synthetic_data(data)
-  mae <-
-    gDRtestData::generateComboMatrixSmall(cell_lines, drugs, e_inf, ec50, hill_coef)
-  
-  test_synthetic_data(original, mae, data)
+  mae <- purrr::quietly(gDRtestData::generateComboMatrixSmall)(
+    cell_lines, drugs, e_inf, ec50, hill_coef, FALSE
+  )
+  expect_length(mae$warnings, 6)
+
+  test_synthetic_data(original, mae$result, data)
 })