+++ rplatform/valid_notes2.R
@@ -1,0 +1,1 @@
+global_valid_notes_l <- list()